Move all summary and plotting to folder make_figures
The scripts that summarize the matches and the ones that plot have become quite many. Also there were a lot of .mat files and pdf's lying around.
I think it's better to just make another folder for that.
I've changed the paths of the scripts to adjust that they are now in
another folder.