Delete rows in first script and save
I delete the NAN rows in find_patents, so that I get the right index number for where patents start.
I then save the deleted rows as well in the cell array pat_ix.

I call this from the second script analyze_patent_text and also delete the rows from the search corpus. It's faster this way, because I don't have to search for NAN twice.