Adjust part1 fdate search
In some of the '
special case' years, the rows are much longer which is why I truncate things. Previously I extracted fdate by going until the end of the line, but this also took a lot of whitespace. So now just exctract the right number of characters.
Also adapt the function that checks the fdate format to show the right patent number for those with strange formatting.