Exclude Pharma and Chemical patents
We collected a number of tech numbers (OCLs) that we wanted to exclude (in our Github issue). I put that in a function that in the directory 'specs'. Now, we also search in our patsearch_results for the technolgoy numbers and save that into automclassData. This will go right into our algorithms.