Also calculate the index for ALL patents:
We use the Silverman concordance to link patents probabilistically to the sectors they are used in. Previously I calculated an automix_use which was the probability of all automation patents to be used in a particular industry, weighted with fractional counts.

Now, I also calculate this for all patents in the year. Why? So that we can compare the two numbers and maybe use the ratio of them.